:wrench: Chore: client-symbol 모듈 구현
<footer>
- 관련: #30
