+++ clients/client-symbol/src/main/java/io/say/better/client/symbol/config/OpenFeignConfig.java
@@ -1,4 +1,4 @@
-package com.saybetter.global.config.web;
+package io.say.better.client.symbol.config;
 
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Configuration;
