+++ clients/client-symbol/src/main/resources/client-symbol.yml
@@ -1,0 +1,25 @@
+## recommend
+recommend:
+  api:
+    base-url: ${RECOMMEND_API_BASE_URL}
+    identity-token: ${RECOMMEND_API_IDENTITY_TOKEN}
+
+spring.cloud.openfeign:
+  client:
+    config:
+      default:
+        connectTimeout: 2100
+        readTimeout: 5000
+        loggerLevel: full
+  compression:
+    response:
+      enabled: false
+  httpclient:
+    max-connections: 2000
+    max-connections-per-route: 500
+
+---
+spring.config.activate.on-profile:
+  - local
+  - local-dev
+  - dev
